put PhoSin qc where it should go
